Installation SQL from running dh.pl
Note that this has created several directories, all with a 1 in, and
each has a __VERSION file in it: this is the special DH version table
used to track the current version of the database. The other file is
just a bunch of CREATE TABLE and what-have-you that will deploy the
schema at this version, i.e. 1